Revert "Adding Watch List Dynamic Module Loading"
