--- watch_list/test1.py
@@ -1,1 +1,0 @@
-print("test 1")
