--- watch_list/test2.py
@@ -1,1 +1,0 @@
-print("test 2")
